# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BuildGraph AI MVP API
  version: 0.2.0
  description: BuildGraph 3주 MVP 공통 API 계약. 상세 기준은 docs/API_CONTRACT.md를 따른다.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/as-tickets:
    post:
      summary: AS 티켓 생성
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AsTicketCreateRequest"
      responses:
        "201":
          description: 생성된 AS 티켓
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsTicketDto"
  /api/as-tickets/{id}:
    get:
      summary: AS 티켓 상세
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PublicId"
      responses:
        "200":
          description: AS 티켓 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsTicketDto"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PublicId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    AsTicketStatus:
      type: string
      enum: [OPEN, ASSIGNED, IN_PROGRESS, RESOLVED, CLOSED, CANCELLED]
...
    AsTicketCreateRequest:
      type: object
      required: [symptom]
      properties:
        logUploadId:
          type: string
          format: uuid
        symptom:
          type: string
    AsTicketDto:
      type: object
      required: [id, status, symptom, causeCandidates, upgradeCandidates]
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/AsTicketStatus"
        symptom:
          type: string
        logUploadId:
          type: string
          format: uuid
        assignedAdminId:
          type: string
          format: uuid
        causeCandidates:
          type: array
          items:
            type: object
            additionalProperties: true
        upgradeCandidates:
          type: array
          items:
            type: object
            additionalProperties: true
        adminNote:
          type: string
        resolvedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...